+++ CHANGELOG.md
@@ -49,6 +49,8 @@ high state of flux, you're at risk of it changing without notice.
   - `ReaderEither`
     - remove `local`, use `Reader`'s `local` instead (@gcanti)
     - remove `getReaderValidation`, use `getApplicativeReaderValidation` and `getAltReaderValidation` instead (@gcanti)
+  - `ReaderTask`
+    - remove `local`, use `Reader`'s `local` instead (@gcanti)
 
 # 2.9.1